clear shader when error

+++ hierarchy/ShaderManager.cpp
@@ -25,7 +25,7 @@ static std::wstring multi_to_wide_winapi(std::string const &src)
 {
     auto const dest_size = ::MultiByteToWideChar(CP_ACP, 0U, src.data(), -1, nullptr, 0U);
     std::vector<wchar_t> dest(dest_size, L'\0');
-    if (::MultiByteToWideChar(CP_ACP, 0U, src.data(), -1, dest.data(), dest.size()) == 0)
+    if (::MultiByteToWideChar(CP_ACP, 0U, src.data(), -1, dest.data(), (int)dest.size()) == 0)
     {
         throw std::system_error{static_cast<int>(::GetLastError()), std::system_category()};
     }
